Custom domains infrastructure
Build the Azure DNS zone and front door to support creating custom
domains for PaaS apps via terraform.
This is based on the front-door "Standard" offering which is not fully
ready in terraform. See progress in: https://github.com/hashicorp/terraform-provider-azurerm/pull/16671

The rest of the set-up is made manually for now and will be added here
once it's ready in terraform.