Use full names for AKS databases
At the moment we use the short name for databases created in Kubernetes,
which ends up with a set of deployments like this:

```
afqts-review-1445-postgres         1/1     1            1           2d22h
afqts-review-1445-redis            1/1     1            1           2d22h
afqts-review-1451-postgres         1/1     1            1           59m
afqts-review-1451-redis            1/1     1            1           59m
apply-for-qts-development-web      1/1     1            1           46m
apply-for-qts-development-worker   1/1     1            1           46m
apply-for-qts-review-1445-web      1/1     1            1           2d22h
apply-for-qts-review-1445-worker   1/1     1            1           2d22h
apply-for-qts-review-1451-web      1/1     1            1           59m
apply-for-qts-review-1451-worker   1/1     1            1           59m
```

This is slightly unclear when first viewing, as the deployments
associated with a particular environment aren't grouped together. Using
the full name should help make this more readable.