Add a document outlining an example Rails deployment
This adds a document which outlines the basic steps necessary to deploy
a Rails application to AKS. It doesn't include the GitHub Actions or
Makefile which I think should be documented in a different place.

This is very much a first start, and probably isn't enough for someone
on their own to start a new application, but it should help us during
the migration and we can iterate it in the future. I also think we
should probably pull this out in to its own GitHub repo template in the
future.