Increase number of pods per node
Production: CPU usage average had been very low: 5.6%. We only hit 60% during some
rare peaks. Since each pod is limited to 1 core and the VMs have 4 cores, it must
have been 2 pods using 50%, plus 10% overhead.
Today's cpu_min is 0.8, which represents 5 pods per VM. We change it to which 0.5
would be 8 pods per VM, so with the same proportion as above, it may create peaks
of about 3 pods. We can afford peaks of 3 pods as this would be 75+10=85% CPU usage.

Test: We observed peaks at 60% CPU but most of the time CPU usage is low as the average
is 17%. And maximum memory usage is 25%. So we half cpu_min to double the number of
pods per node, which would only impact apps performance during peak time as CPU will
be slightly constrained, but still acceptable for test.