Prevent statuscake SSL checks from following redirects
When a domain is set up with a redirect, the StatusCake SSL check was
following the redirect and got the data of the end website certificate.
For instance, https://professional-development-for-teachers-leaders.education.gov.uk/
redirects to https://www.gov.uk/guidance/national-professional-qualification-npq-courses.
So the check was on www.gov.uk instead of professional-evelopment-for-teachers-leaders.education.gov.uk.

Remove the follow redirect option as the default value is false

Checks for each domain should be added explicitly if required.