Ignore changes to public_network_access_enabled
Related to issue https://github.com/hashicorp/terraform-provider-azurerm/issues/26098

We must ignore changes to the argument as it becomes enforced from
azurerm version v3.105.0.
This will allow us to upgrade azurerm in all the repositories. Then we
will be able to set public_network_access_enabled to false and remove
this ignore_changes.