Fix postgres database password issues with $$
We set the postgres password in the container using a shell environment
variable. If the variable contains a sequence of 2 dollar signs ($$),
the actual password may actually have only one ($). Since the webapp is
still configured with $$, they don't match and authentication fails.

This commit replaces all sequences of multiple dollar signs with a
single one.