Repeat London as StatusCake region
Each region has multiple StatusCake servers. If it is repeated, the
confirmations will use another server in the same region. We want the
London region as most of our users are in UK. Focusing on it means we
may not alert if a service is down for London but available for another
country, as may happen with the front door CDN