Allow 1 character custom domains length
The name of the custom domain in front door is the same as the subdomain.
Subdomain with 1 character is allowed but not the custom domain name.

Add a fix to create a longer name when the subdomain is 1 character.

This is required as the length of "<domain>.<zone>" cannot exceed 64
characters and some subdomains must be 1 character to fit in.